requirements.txt
streamlit
numpy
pandas
scikit-learn
matplotlib
statsmodels
fpdf
python-telegram-bot
